Add unit tests for mappers

See merge request mobile/pax-app-core/app-sizer!27

+++ clt/src/main/kotlin/com/grab/sizer/CltInputProvider.kt
@@ -61,7 +61,7 @@ class CltInputProvider constructor(
 
     override fun provideTeamMappingFile(): File? = config.projectInput.ownerMappingFile
 
-    override fun provideLargeFileThreshold(): Int = config.projectInput.largeFileThreshold
+    override fun provideLargeFileThreshold(): Long = config.projectInput.largeFileThreshold
 }
